+++ viewer/src/views/layout/left/LayoutProposition.vue
@@ -119,6 +119,7 @@ const propositions = computed<Record<string, number>>(() => {
     // Tags count from current search
     extractDistinctItems(searchFilters)
       .flatMap(item => item.tags)
+      .flatMap(tag => galleryStore.galleryIndex?.tags[tag])
       .map(rightmost)
       .filter(rawTag => props.tagsIndex[rawTag] && !searchFilters.find(search => search.tag === rawTag))
       .forEach(rawTag => (propositions[rawTag] = (propositions[rawTag] ?? 0) + 1));