+++ changelog.md
@@ -4,6 +4,11 @@ This file lists notable changes that have been made to the application on each
 release. Releases are tracked and referred to using git tags.
 
 
+## Upcoming changes for next version
+- Improvements:
+  - The size of the gallery `index.json` file has been reduced.
+
+
 ## v2.1 - 2022-09-04
 - New features:
   - Add support for Markdown-formatted files, which are now rendered.